backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.0.0
Markdown==3.4.1
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pytz==2022.7.1
requests==2.28.2
shortuuid==1.0.11
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
zipp==3.12.0
